+++ Classes_Metodos/src/Heranca/Teste.java
@@ -1,0 +1,24 @@
+/*
+ * To change this license header, choose License Headers in Project Properties.
+ * To change this template file, choose Tools | Templates
+ * and open the template in the editor.
+ */
+package Herenca;
+
+/**
+ *
+ * @author MP
+ */
+public class Teste {
+
+    public static void main(String[] args) {
+        Mae mae = new Mae();
+        mae.setNome("Mae");
+        mae.andar();
+
+        Filha filha = new Filha();
+        filha.setNome("Filha");
+        filha.andar();
+   
+    }
+}
